+++ spec/adhearsion/call_controller/record_spec.rb
@@ -14,16 +14,17 @@ module Adhearsion
             :max_duration => 5000
           }
         end
-        let(:component) { ::Punchblock::Component::Record.new options }
-        let(:response)  { ::Punchblock::Event::Complete.new }
+        let(:component) { Punchblock::Component::Record.new options }
+        let(:response)  { Punchblock::Event::Complete.new }
 
         describe "with :async => true and an :on_complete callback" do
           before do
             component
-            flexmock(::Punchblock::Component::Record).should_receive(:new).once.with(options).and_return component
+            flexmock(Punchblock::Component::Record).should_receive(:new).once.with(options).and_return component
             expect_message_waiting_for_response component
             @rec = Queue.new
             subject.record(options.merge(async: true)) { |rec| @rec.push rec }
+            component.request!
             component.execute!
           end
 
@@ -39,7 +40,7 @@ module Adhearsion
           before do
             TestException = Class.new StandardError
             component
-            flexmock(::Punchblock::Component::Record).should_receive(:new).once.with({}).and_return component
+            flexmock(Punchblock::Component::Record).should_receive(:new).once.with({}).and_return component
           end
 
           it "should pass the exception to the events system" do
@@ -62,7 +63,7 @@ module Adhearsion
         describe "with :async => false" do
           before do
             component
-            flexmock(::Punchblock::Component::Record).should_receive(:new).once.with(options).and_return component
+            flexmock(Punchblock::Component::Record).should_receive(:new).once.with(options).and_return component
             expect_component_execution component
             @rec = Queue.new
             subject.record(options.merge(:async => false)) { |rec| @rec.push rec }
@@ -78,9 +79,8 @@ module Adhearsion
           end
         end
 
-        
         describe "with :interruptible => false" do
-          let(:input_component) { ::Punchblock::Component::Input.new }
+          let(:input_component) { Punchblock::Component::Input.new }
           it "does not use an Input component" do
             subject.should_receive(:execute_component_and_await_completion).once.with(component)
             subject.should_receive(:write_and_await_response).never.with(input_component)
@@ -89,7 +89,7 @@ module Adhearsion
         end
 
         describe "with :interruptible => true" do
-          let(:input_component) { ::Punchblock::Component::Input.new }
+          let(:input_component) { Punchblock::Component::Input.new }
           it "stops the recording" do
             flexmock(Punchblock::Event::Complete).new_instances.should_receive(:reason => flexmock(:name => :input))
 
@@ -106,14 +106,14 @@ module Adhearsion
             flexmock(Punchblock::Component::Record).new_instances.should_receive(:stop!)
             subject.should_receive(:execute_component_and_await_completion).once.with(component)
             subject.record(options.merge(:async => false, :interruptible => true)) { |rec| @rec.push rec }
-          end 
+          end
 
         end
 
         describe "check for the return value" do
           it "returns a Record component" do
             component
-            flexmock(::Punchblock::Component::Record).should_receive(:new).once.with(options).and_return component
+            flexmock(Punchblock::Component::Record).should_receive(:new).once.with(options).and_return component
             expect_component_execution component
             subject.record(options.merge(:async => false)).should == component
             component.request!